Rename and consolidate Subthoughts components and hooks.
