+++ src/components/DropCliff.tsx
@@ -2,16 +2,16 @@ import classNames from 'classnames'
 import React from 'react'
 import { ConnectDropTarget } from 'react-dnd'
 import { useSelector } from 'react-redux'
-import DropThoughtZone from '../../@types/DropThoughtZone'
-import SimplePath from '../../@types/SimplePath'
-import State from '../../@types/State'
-import globals from '../../globals'
-import getThoughtById from '../../selectors/getThoughtById'
-import head from '../../util/head'
-import strip from '../../util/strip'
-import DragAndDropThought from '../DragAndDropThought'
-import useDropHoverColor from './useDropHoverColor'
-import useHoveringPath from './useHoveringPath'
+import DropThoughtZone from '../@types/DropThoughtZone'
+import SimplePath from '../@types/SimplePath'
+import State from '../@types/State'
+import globals from '../globals'
+import useDropHoverColor from '../hooks/useDropHoverColor'
+import useHoveringPath from '../hooks/useHoveringPath'
+import getThoughtById from '../selectors/getThoughtById'
+import head from '../util/head'
+import strip from '../util/strip'
+import DragAndDropThought from './DragAndDropThought'
 
 /** A drop target for after the hidden parent at a cliff. This is needed because the Thought will be hidden/shimmed so DragAndDropThought will not be rendered. */
 const SubthoughtsDropCliff = ({
