+++ src/components/DropEmpty.tsx
@@ -2,18 +2,18 @@ import classNames from 'classnames'
 import React from 'react'
 import { ConnectDropTarget } from 'react-dnd'
 import { useSelector } from 'react-redux'
-import DropThoughtZone from '../../@types/DropThoughtZone'
-import SimplePath from '../../@types/SimplePath'
-import State from '../../@types/State'
-import globals from '../../globals'
-import getThoughtById from '../../selectors/getThoughtById'
-import equalPath from '../../util/equalPath'
-import head from '../../util/head'
-import isDivider from '../../util/isDivider'
-import strip from '../../util/strip'
-import DragAndDropSubthoughts from '../DragAndDropSubthoughts'
-import useDropHoverColor from './useDropHoverColor'
-import useHoveringPath from './useHoveringPath'
+import DropThoughtZone from '../@types/DropThoughtZone'
+import SimplePath from '../@types/SimplePath'
+import State from '../@types/State'
+import globals from '../globals'
+import useDropHoverColor from '../hooks/useDropHoverColor'
+import useHoveringPath from '../hooks/useHoveringPath'
+import getThoughtById from '../selectors/getThoughtById'
+import equalPath from '../util/equalPath'
+import head from '../util/head'
+import isDivider from '../util/isDivider'
+import strip from '../util/strip'
+import DragAndDropSubthoughts from './DragAndDropSubthoughts'
 
 /** A drop target when there are no children or the thought is collapsed. The drop-hover components are ThoughtDropHover, SubthoughtsDropEnd, and SubthoughtsDropEmpty. */
 const SubthoughtsDropEmpty = ({
