+++ src/components/DropEnd.tsx
@@ -2,25 +2,25 @@ import classNames from 'classnames'
 import React, { FC } from 'react'
 import { ConnectDropTarget } from 'react-dnd'
 import { useSelector } from 'react-redux'
-import DropThoughtZone from '../../@types/DropThoughtZone'
-import SimplePath from '../../@types/SimplePath'
-import State from '../../@types/State'
-import { isTouch } from '../../browser'
-import { ID } from '../../constants'
-import globals from '../../globals'
-import { getChildrenSorted } from '../../selectors/getChildren'
-import getSortPreference from '../../selectors/getSortPreference'
-import getThoughtById from '../../selectors/getThoughtById'
-import rootedParentOf from '../../selectors/rootedParentOf'
-import { compareReasonable } from '../../util/compareThought'
-import equalPath from '../../util/equalPath'
-import head from '../../util/head'
-import headId from '../../util/headId'
-import isRoot from '../../util/isRoot'
-import strip from '../../util/strip'
-import DragAndDropSubthoughts from '../DragAndDropSubthoughts'
-import useDropHoverColor from './useDropHoverColor'
-import useHoveringPath from './useHoveringPath'
+import DropThoughtZone from '../@types/DropThoughtZone'
+import SimplePath from '../@types/SimplePath'
+import State from '../@types/State'
+import { isTouch } from '../browser'
+import { ID } from '../constants'
+import globals from '../globals'
+import useDropHoverColor from '../hooks/useDropHoverColor'
+import useHoveringPath from '../hooks/useHoveringPath'
+import { getChildrenSorted } from '../selectors/getChildren'
+import getSortPreference from '../selectors/getSortPreference'
+import getThoughtById from '../selectors/getThoughtById'
+import rootedParentOf from '../selectors/rootedParentOf'
+import { compareReasonable } from '../util/compareThought'
+import equalPath from '../util/equalPath'
+import head from '../util/head'
+import headId from '../util/headId'
+import isRoot from '../util/isRoot'
+import strip from '../util/strip'
+import DragAndDropSubthoughts from './DragAndDropSubthoughts'
 
 /** A container fragment that only renders its children when dragInProgress is true. Useful for short circuiting child components with more expensive selectors. */
 const DragOnly: FC = ({ children }) => {
