+++ src/components/Subthoughts.native.tsx
@@ -13,6 +13,7 @@ import ThoughtContext from '../@types/ThoughtContext'
 import ThoughtId from '../@types/ThoughtId'
 import { MAX_DEPTH, MAX_DISTANCE_FROM_CURSOR, ViewMode } from '../constants'
 import globals from '../globals'
+import useZoom from '../hooks/useZoom'
 import appendChildPath from '../selectors/appendChildPath'
 import attribute from '../selectors/attribute'
 import childIdsToThoughts from '../selectors/childIdsToThoughts'
@@ -53,7 +54,6 @@ import parseLet from '../util/parseLet'
 import pathToContext from '../util/pathToContext'
 import { safeRefMerge } from '../util/safeRefMerge'
 import GestureDiagram from './GestureDiagram'
-import useZoom from './Subthoughts/useZoom'
 import { Text } from './Text.native'
 import Thought from './Thought'
