+++ src/components/Subthoughts.tsx
@@ -17,6 +17,7 @@ import dragInProgress from '../action-creators/dragInProgress'
 import { isTouch } from '../browser'
 import { HOME_TOKEN, MAX_DEPTH, MAX_DISTANCE_FROM_CURSOR } from '../constants'
 import globals from '../globals'
+import useZoom from '../hooks/useZoom'
 import appendChildPath from '../selectors/appendChildPath'
 import attribute from '../selectors/attribute'
 import childIdsToThoughts from '../selectors/childIdsToThoughts'
@@ -59,7 +60,6 @@ import pathToContext from '../util/pathToContext'
 import safeRefMerge from '../util/safeRefMerge'
 import unroot from '../util/unroot'
 import GestureDiagram from './GestureDiagram'
-import useZoom from './Subthoughts/useZoom'
 import Thought from './Thought'
 
 /** The type of the exported Subthoughts. */
