+++ src/components/ThoughtDropHover.tsx
@@ -1,19 +1,19 @@
 import React from 'react'
 import { useSelector } from 'react-redux'
-import DropThoughtZone from '../../@types/DropThoughtZone'
-import SimplePath from '../../@types/SimplePath'
-import State from '../../@types/State'
-import ThoughtId from '../../@types/ThoughtId'
-import globals from '../../globals'
-import getSortPreference from '../../selectors/getSortPreference'
-import getThoughtById from '../../selectors/getThoughtById'
-import rootedParentOf from '../../selectors/rootedParentOf'
-import { compareReasonable } from '../../util/compareThought'
-import equalPath from '../../util/equalPath'
-import head from '../../util/head'
-import headId from '../../util/headId'
-import parentOf from '../../util/parentOf'
-import useDropHoverColor from '../Subthoughts/useDropHoverColor'
+import DropThoughtZone from '../@types/DropThoughtZone'
+import SimplePath from '../@types/SimplePath'
+import State from '../@types/State'
+import ThoughtId from '../@types/ThoughtId'
+import globals from '../globals'
+import useDropHoverColor from '../hooks/useDropHoverColor'
+import getSortPreference from '../selectors/getSortPreference'
+import getThoughtById from '../selectors/getThoughtById'
+import rootedParentOf from '../selectors/rootedParentOf'
+import { compareReasonable } from '../util/compareThought'
+import equalPath from '../util/equalPath'
+import head from '../util/head'
+import headId from '../util/headId'
+import parentOf from '../util/parentOf'
 
 /** A drop-hover element that is rendered during drag-and-drop when it is possible to drop in a ThoughtDrop zone (next to a Thought). The drop-hover components are ThoughtDropHover, SubthoughtsDropEnd, and SubthoughtsDropEmpty. */
 const ThoughtDropHover = ({
