+++ src/hooks/useHideBullet.ts
