+++ src/hooks/useHoveringPath.ts
@@ -1,9 +1,9 @@
 import { useEffect } from 'react'
 import { useDispatch } from 'react-redux'
-import DragThoughtZone from '../../@types/DragThoughtZone'
-import DropThoughtZone from '../../@types/DropThoughtZone'
-import Path from '../../@types/Path'
-import dragInProgress from '../../action-creators/dragInProgress'
+import DragThoughtZone from '../@types/DragThoughtZone'
+import DropThoughtZone from '../@types/DropThoughtZone'
+import Path from '../@types/Path'
+import dragInProgress from '../action-creators/dragInProgress'
 
 /** Updates state.hoveringPath when isHovering becomes true. Used to determine the parent of state.hoveringPath to be highlighted (via isChildHovering in the Thought components), expandHoverBottom/Top, and the visibility of drop-hover elements. */
 const useHoveringPath = (path: Path, isHovering: boolean, hoverZone: DropThoughtZone) => {
