+++ src/hooks/useThoughtStyleContainer.ts
