+++ src/hooks/useZoom.ts
@@ -1,14 +1,14 @@
 import { useSelector } from 'react-redux'
-import LazyEnv from '../../@types/LazyEnv'
-import SimplePath from '../../@types/SimplePath'
-import State from '../../@types/State'
-import attributeEquals from '../../selectors/attributeEquals'
-import findDescendant from '../../selectors/findDescendant'
-import rootedParentOf from '../../selectors/rootedParentOf'
-import equalPath from '../../util/equalPath'
-import findFirstEnvContextWithZoom from '../../util/findFirstEnvContextWithZoom'
-import head from '../../util/head'
-import parentOf from '../../util/parentOf'
+import LazyEnv from '../@types/LazyEnv'
+import SimplePath from '../@types/SimplePath'
+import State from '../@types/State'
+import attributeEquals from '../selectors/attributeEquals'
+import findDescendant from '../selectors/findDescendant'
+import rootedParentOf from '../selectors/rootedParentOf'
+import equalPath from '../util/equalPath'
+import findFirstEnvContextWithZoom from '../util/findFirstEnvContextWithZoom'
+import head from '../util/head'
+import parentOf from '../util/parentOf'
 
 /** Calculates the cursor zoom after initial render for better performance. */
 const useZoom = ({
